+++ Package.swift
@@ -11,13 +11,9 @@ let package = Package(
         .library(
             name: "SDOSAlamofire",
             targets: ["SDOSAlamofire"]),
-        .library(
-            name: "SDOSAlamofireJSONAPI",
-            targets: ["SDOSAlamofireJSONAPI"])
     ],
     dependencies: [
         .package(url: "https://github.com/Alamofire/Alamofire.git", .upToNextMajor(from: "5.2.0")),
-        .package(url: "https://github.com/SDOSLabs/Japx.git", .branch("feature/spm")),
         .package(url: "https://github.com/SDOSLabs/SDOSKeyedCodable.git", .branch("feature/spm")),
         .package(url: "https://github.com/SDOSLabs/SDOSSwiftExtension.git", .branch("feature/spm"))
         
@@ -31,13 +27,5 @@ let package = Package(
                 "SDOSSwiftExtension"
             ],
             path: "src/Classes/Core"),
-        .target(
-            name: "SDOSAlamofireJSONAPI",
-            dependencies: [
-                "SDOSAlamofire",
-                .product(name: "JapxCodable", package: "Japx")
-            ],
-            path: "src/Classes/JSONAPI")
-        
     ]
 )